feat: custom theme loading
Adds a custom theme nickvd, which at the moment is just a copy
of robbyrussels theme from ohmyzsh. In the future I might tweak
to be more personalised. It also adds in a kali linux theme, which
at the moment is just a copy from a default kali linux install. Again
I might tweak this in the future.

It also adds some logic around theme selection. On kali linux I
prefer using the kali-linux theme. I mostly use kali linux in a
vm. Having a different theme for this distro means that when I
ssh into those boxes I can clearly see when I'm in a VM. Or when
I'm executing commmands on my host.