fix: tmux sessionizer breaks when folder does not exist
Sessionizer fails when a folder passed to find doesn not exist. Find
then returns a NZEC which due to my bash defaults exits the progrom.
This means that I never run the code that switches or creates the new
session. This patch also supresses any errors find writes to stderr when
a folder doens't exist. We just don't care about these errors and they
are noisy.

The reason for this change is that it is technically possible to pass a
non existing folder to -w. But also my defaults are ~/Projects and
~/Work. I don't have a work folder on my personal computer. I rather
have find silently fail and show a list of empty folders. This is what
did already and indicates to the user that something went wrong.